feat(apps): parse compose through the shared lnvps-compose package
Replaces src/utils/app-compose.ts (js-yaml, config-only, lenient by
accident) with the shared parser also used by the operator's Rust
source. DEPLOYMENT_NAME_RE and defaultConfigValues move to
src/utils/deployment-form.ts — form-state helpers, not compose grammar.

lnvps-compose installs as a git dependency; its prepare script builds
dist/ on install, which bun blocks by default for git deps, hence
trustedDependencies. Verified against a real production install and
a real Docker build/run, not just a local one — web#85 was exactly
this class of miss.
