+++ package.json
@@ -37,7 +37,7 @@
     "classnames": "^2.5.1",
     "dns-over-http-resolver": "^3.0.16",
     "iso-3166-1": "^2.1.1",
-    "js-yaml": "^4.1.1",
+    "lnvps-compose": "github:LNVPS/lnvps-compose#v0.1.0",
     "marked": "^15.0.7",
     "qr-code-styling": "^1.8.4",
     "react": "^19.1.0",
@@ -53,7 +53,6 @@
     "@types/bun": "^1.3.10",
     "@types/compression": "^1.8.1",
     "@types/express": "^5.0.6",
-    "@types/js-yaml": "^4.0.9",
     "@types/react": "^19.1.0",
     "@types/react-dom": "^19.1.0",
     "@vitejs/plugin-react": "^5.1.2",
@@ -76,5 +75,8 @@
     "lodash": "^4.17.23",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
-  }
+  },
+  "trustedDependencies": [
+    "lnvps-compose"
+  ]
 }
